+++ src/main/scala/au/com/agl/arc/extract/HTTPExtract.scala
@@ -6,15 +6,15 @@ import scala.collection.JavaConverters._
 
 import org.apache.http.client.methods.{HttpGet, HttpPost}
 import org.apache.http.impl.client.HttpClientBuilder
+import org.apache.http.impl.client.HttpClients
+import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
 import org.apache.http.entity.StringEntity
 import org.apache.http.impl.client.LaxRedirectStrategy
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.scala.DefaultScalaModule
-
 import org.apache.spark.sql._
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.catalyst.ScalaReflection
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types._
 import org.apache.spark.storage.StorageLevel
 
 import scala.io.Source
@@ -23,7 +23,20 @@ import au.com.agl.arc.api._
 import au.com.agl.arc.api.API._
 import au.com.agl.arc.util._
 
+case class RequestResponse(
+    uri: String,
+    statusCode: Int,
+    reasonPhrase: String,
+    contentType: String,
+    body: String
+)
+
+
 object HTTPExtract {
+
+  /** Phantom Type to enable compiler to find the encoder we want
+    */
+  type RequestResponseRow = Row
 
   def extract(extract: HTTPExtract)(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger): Option[DataFrame] = {
     import spark.implicits._
@@ -35,62 +48,130 @@ object HTTPExtract {
     val stageDetail = new java.util.HashMap[String, Object]()
     stageDetail.put("type", extract.getType)
     stageDetail.put("name", extract.name)
-    stageDetail.put("uri", extract.uri.toString)  
     stageDetail.put("outputView", extract.outputView)  
     stageDetail.put("persist", Boolean.valueOf(extract.persist))
     stageDetail.put("method", method)
     stageDetail.put("headers", maskedHeaders.asJava)
+
+    val inputValue = extract.input match {
+      case Right(uri) => uri.toString
+      case Left(view) => view
+    }
+
+    stageDetail.put("input", inputValue)  
 
     logger.info()
       .field("event", "enter")
       .map("stage", stageDetail)      
       .log()  
 
-    val uri = extract.uri.toString
+    // create a StructType schema for RequestResponse
+    val typedSchema = ScalaReflection.schemaFor[RequestResponse].dataType.asInstanceOf[StructType]      
 
-    val client = HttpClientBuilder.create.setRedirectStrategy(new LaxRedirectStrategy()).build
+    /** Create a dynamic RowEncoder from the provided schema. We use the phantom
+      * TypeRow type to enable implicit resolution to find our encoder.
+      */    
+    implicit val typedEncoder: Encoder[RequestResponseRow] = org.apache.spark.sql.catalyst.encoders.RowEncoder(typedSchema)
 
-    val request = method match {
-      case "GET" => new HttpGet(uri)
-      case "POST" => { 
-        val post = new HttpPost(uri)
-        for (body <- extract.body) {
-          post.setEntity(new StringEntity(body))
+    val responses = try {
+      val df = extract.input match {
+        case Right(uri) => Seq(uri.toString).toDF("value")
+        case Left(view) => spark.table(view)
+      }
+
+      df.mapPartitions[RequestResponseRow] { partition: Iterator[Row] => 
+        val poolingHttpClientConnectionManager = new PoolingHttpClientConnectionManager()
+        poolingHttpClientConnectionManager.setMaxTotal(50)
+        val httpClient = HttpClients.custom()
+          .setRedirectStrategy(new LaxRedirectStrategy())
+          .setConnectionManager(poolingHttpClientConnectionManager)
+          .build()
+
+        // we are using a BufferedIterator so we can 'peek' at the first row to get column types without advancing the iterator
+        // meaning we don't have to keep finding fieldIndex and dataType for each row (inefficient as they will not change)
+        val bufferedPartition = partition.buffered
+        val uriFieldIndex = bufferedPartition.hasNext match {
+          case true => bufferedPartition.head.fieldIndex("value")
+          case false => 0
         }
-        post 
+
+        bufferedPartition.map[RequestResponseRow] { row: Row =>
+          val uri = row.getString(uriFieldIndex)
+
+          val request = method match {
+            case "GET" => new HttpGet(uri)
+            case "POST" => { 
+              val post = new HttpPost(uri)
+              for (body <- extract.body) {
+                post.setEntity(new StringEntity(body))
+              }
+              post 
+            }
+          }
+
+          // add headers
+          for ((k,v) <- extract.headers) {
+            request.addHeader(k,v) 
+          }
+
+          try {
+            // send the request
+            val response = httpClient.execute(request)
+
+            // read and close response
+            val body = response.getEntity.getContentLength match {
+              case 0 => None
+              case _ => Option(Source.fromInputStream(response.getEntity.getContent).mkString)
+            }
+            response.close 
+
+            // cast to a RequestResponseRow to fit the Dataset map method requirements
+            val result = Seq(uri, response.getStatusLine.getStatusCode, response.getStatusLine.getReasonPhrase, Option(response.getEntity.getContentType).map(_.toString).orNull, body.orNull)
+            Row.fromSeq(result).asInstanceOf[RequestResponseRow]
+          } finally {
+            request.releaseConnection
+          }
+        }
+      }
+    } catch {
+      case e: Exception => throw new Exception(e) with DetailException {
+        override val detail = stageDetail          
       }
     }
-    // add headers
-    for ((k,v) <- extract.headers) {
-      request.addHeader(k,v) 
-    }
 
-    // send the request
-    val response = client.execute(request)
+    val df = responses.toDF
 
-    val responseMap = new java.util.HashMap[String, Object]()
-    responseMap.put("statusCode", new java.lang.Integer(response.getStatusLine.getStatusCode))
-    responseMap.put("reasonPhrase", response.getStatusLine.getReasonPhrase)      
-    stageDetail.put("response", responseMap)   
+    val distinctReponses = responses
+      .groupBy($"statusCode", $"reasonPhrase")
+      .agg(collect_list($"body").as("body"), count("*").as("count"))
+
+    // execute the requests and return a new dataset of distinct response codes
+    distinctReponses.cache.count
+
+    // response logging 
+    // limited to 50 top response codes (by count descending) to protect against log flooding
+    val responseMap = new java.util.HashMap[String, Object]()      
+    distinctReponses.sort($"count".desc).limit(50).collect.foreach( response => {
+      val colMap = new java.util.HashMap[String, Object]()
+      colMap.put("body", response.getList(2).toArray.slice(0, 10).distinct)
+      colMap.put("reasonPhrase", response.getString(1))
+      colMap.put("count", Long.valueOf(response.getLong(3)))
+      responseMap.put(response.getInt(0).toString, colMap)
+    })
+    stageDetail.put("responses", responseMap)    
 
     // verify status code is correct
     val validStatusCodes = extract.validStatusCodes match {
       case Some(value) => value
       case None => 200 :: 201 :: 202 :: Nil
     }
-    if (!validStatusCodes.contains(response.getStatusLine.getStatusCode)) {
-      response.close    
-      throw new Exception(s"""HTTPExtract expects a response StatusCode in [${validStatusCodes.mkString(", ")}] but server responded with ${response.getStatusLine.getStatusCode} (${response.getStatusLine.getReasonPhrase}).""") with DetailException {
-        override val detail = stageDetail
+    if (!(distinctReponses.map(d => d.getInt(0)).collect forall (validStatusCodes contains _))) {
+      val responseMessages = distinctReponses.map(response => s"${response.getLong(3)} reponses ${response.getInt(0)} (${response.getString(1)})").collect.mkString(", ")
+
+      throw new Exception(s"""HTTPExtract expects all response StatusCode(s) in [${validStatusCodes.mkString(", ")}] but server responded with [${responseMessages}].""") with DetailException {
+        override val detail = stageDetail          
       }
-    }    
-
-    // read and close response
-    val responseEntity = response.getEntity.getContent
-    val body = Source.fromInputStream(responseEntity).mkString.replace("\n", "")
-    response.close
-
-    val df = spark.sparkContext.parallelize(Array(body)).toDF
+    }   
     
     // repartition to distribute rows evenly
     val repartitionedDF = extract.partitionBy match {